requirements.txt
pandas
nltk
flask
flask-socketio
python-eventlet
requests